fix for buffer distance algorithm
i found that it was necessary to specify a greater distance than should
be necessary in order to retrieve buffer results when ShowBeyondBuffer
is set to 'false' for a feature layer.

example: searching the address "5323 Velvent Bent Ct Naperville IL
60564" with a distance of 10 only returns one railroad station 7.11
miles away, even though the other is 7.8 miles away.

after switching to use WGS84 as buffer spatial reference, a distance of
8 in the config.js causes both to be displayed.